cloud-storage/aws-lambda-package/requirements.txt
aws-psycopg2==1.3.8
psycopg2-binary==2.9.9